[move-to-yul] Cleanup
This PR is basically a functional no-op but cleans up the code structure a bit before it goes to much out of hand. Specifically, the monolithic module `generator.rs` has become too large.

- Move Context into its own module `context.rs`
- Move function and bytecode generation out of `generator.rs` into a new module `functions.rs`
- Move attributes into new module `attributes`
- Renames some functions

Hopefully no one has significant local work in progress because rebase against this might be ugly.

Closes: #72
