+++ language/evm/move-to-yul/tests/NoGenericCallable.exp
@@ -1,12 +1,3 @@
-!! Move-To-Yul Diagnostics:
- error: #[callable] or #[create] functions cannot be generic
-  ┌─ tests/NoGenericCallable.move:5:2
-  │
-5 │     fun f<T>(x: u64): u64 { x }
-  │     ^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-
-
 /* =======================================
  * Generated by Move-To-Yul compiler v0.0
  * ======================================= */
@@ -41,30 +32,6 @@ object "A2_M" {
                 default {}
             }
             $Abort(97)
-            function abi_decode_tuple_$u64$(headStart, dataEnd) -> value_0 {
-                if slt(sub(dataEnd, headStart), 32) { $Abort(96) }
-                {
-                    let offset := 0
-                    value_0 := abi_decode_u64(add(headStart, offset), dataEnd)
-                }
-            }
-            function abi_decode_u64(offset, end) -> value {
-                value := calldataload(offset)
-                validator_u64(value)
-            }
-            function abi_encode_tuple_$u64$(headStart ,value_0) -> tail {
-                tail := add(headStart, 32)
-                abi_encode_u64(value_0, add(headStart, 0))
-            }
-            function abi_encode_u64(value, pos) {
-                mstore(pos, cleanup_u64(value))
-            }
-            function cleanup_u64(value) -> cleaned {
-                cleaned := and(value, 0xffffffffffffffff)
-            }
-            function validator_u64(value) {
-                if iszero(eq(value, cleanup_u64(value))) { $Abort(95) }
-            }
 
             // Dummy calls to reference callables for Yul optimizer
             let $dummy0 := A2_M_f(sload(100))
@@ -76,6 +43,30 @@ object "A2_M" {
                 $result := x
             }
 
+            function abi_encode_tuple_$u64$(headStart ,value_0) -> tail {
+                tail := add(headStart, 32)
+                abi_encode_u64(value_0, add(headStart, 0))
+            }
+            function abi_encode_u64(value, pos) {
+                mstore(pos, cleanup_u64(value))
+            }
+            function cleanup_u64(value) -> cleaned {
+                cleaned := and(value, 0xffffffffffffffff)
+            }
+            function abi_decode_tuple_$u64$(headStart, dataEnd) -> value_0 {
+                if slt(sub(dataEnd, headStart), 32) { $Abort(96) }
+                {
+                    let offset := 0
+                    value_0 := abi_decode_u64(add(headStart, offset), dataEnd)
+                }
+            }
+            function abi_decode_u64(offset, end) -> value {
+                value := calldataload(offset)
+                validator_u64(value)
+            }
+            function validator_u64(value) {
+                if iszero(eq(value, cleanup_u64(value))) { $Abort(95) }
+            }
             function $Abort(code) {
                 revert(0, 0) // TODO: convention to store code
             }
@@ -85,3 +76,11 @@ object "A2_M" {
         }
     }
 }
+
+
+!! Move-To-Yul Diagnostics:
+ error: #[callable] or #[create] functions cannot be generic
+  ┌─ tests/NoGenericCallable.move:5:2
+  │
+5 │     fun f<T>(x: u64): u64 { x }
+  │     ^^^^^^^^^^^^^^^^^^^^^^^^^^^
