+++ language/evm/move-to-yul/tests/fallback-receive-test/expect-failure/FallbackMultipleParams.exp
@@ -1,14 +1,3 @@
-!! Move-To-Yul Diagnostics:
- error: fallback function must have at most 1 parameter and 1 return value
-  ┌─ tests/fallback-receive-test/expect-failure/FallbackMultipleParams.move:5:2
-  │
-5 │ ╭     fun fallback(x: u64, y: u64): u64 {
-6 │ │         x + y
-7 │ │     }
-  │ ╰─────^
-
-
-
 /* =======================================
  * Generated by Move-To-Yul compiler v0.0
  * ======================================= */
@@ -64,3 +53,13 @@ object "A2_M" {
         }
     }
 }
+
+
+!! Move-To-Yul Diagnostics:
+ error: fallback function must have at most 1 parameter and 1 return value
+  ┌─ tests/fallback-receive-test/expect-failure/FallbackMultipleParams.move:5:2
+  │
+5 │ ╭     fun fallback(x: u64, y: u64): u64 {
+6 │ │         x + y
+7 │ │     }
+  │ ╰─────^
