+++ language/evm/move-to-yul/tests/fallback-receive-test/expect-failure/FallbackReciveFailure.exp
@@ -1,54 +1,3 @@
-!! Move-To-Yul Diagnostics:
- error: multiple #[receive] functions
-   ┌─ tests/fallback-receive-test/expect-failure/FallbackReciveFailure.move:10:2
-   │
-10 │ ╭     fun receive(x: u64): u64 {
-11 │ │         x
-12 │ │     }
-   │ ╰─────^
-
-error: receive function must be payable
-   ┌─ tests/fallback-receive-test/expect-failure/FallbackReciveFailure.move:10:2
-   │
-10 │ ╭     fun receive(x: u64): u64 {
-11 │ │         x
-12 │ │     }
-   │ ╰─────^
-
-error: receive function must not be a fallback or callable function
-   ┌─ tests/fallback-receive-test/expect-failure/FallbackReciveFailure.move:10:2
-   │
-10 │ ╭     fun receive(x: u64): u64 {
-11 │ │         x
-12 │ │     }
-   │ ╰─────^
-
-error: receive function must not have parameters
-   ┌─ tests/fallback-receive-test/expect-failure/FallbackReciveFailure.move:10:2
-   │
-10 │ ╭     fun receive(x: u64): u64 {
-11 │ │         x
-12 │ │     }
-   │ ╰─────^
-
-error: multiple #[fallback] functions
-   ┌─ tests/fallback-receive-test/expect-failure/FallbackReciveFailure.move:10:2
-   │
-10 │ ╭     fun receive(x: u64): u64 {
-11 │ │         x
-12 │ │     }
-   │ ╰─────^
-
-error: fallback function must not be a callable function
-   ┌─ tests/fallback-receive-test/expect-failure/FallbackReciveFailure.move:10:2
-   │
-10 │ ╭     fun receive(x: u64): u64 {
-11 │ │         x
-12 │ │     }
-   │ ╰─────^
-
-
-
 /* =======================================
  * Generated by Move-To-Yul compiler v0.0
  * ======================================= */
@@ -102,30 +51,6 @@ object "A2_M" {
             }
             let retval := A2_M_receive(0, calldatasize()) stop()
             return(add(retval, 0x20), mload(retval))
-            function abi_decode_tuple_$u64$(headStart, dataEnd) -> value_0 {
-                if slt(sub(dataEnd, headStart), 32) { $Abort(96) }
-                {
-                    let offset := 0
-                    value_0 := abi_decode_u64(add(headStart, offset), dataEnd)
-                }
-            }
-            function abi_decode_u64(offset, end) -> value {
-                value := calldataload(offset)
-                validator_u64(value)
-            }
-            function abi_encode_tuple_$u64$(headStart ,value_0) -> tail {
-                tail := add(headStart, 32)
-                abi_encode_u64(value_0, add(headStart, 0))
-            }
-            function abi_encode_u64(value, pos) {
-                mstore(pos, cleanup_u64(value))
-            }
-            function cleanup_u64(value) -> cleaned {
-                cleaned := and(value, 0xffffffffffffffff)
-            }
-            function validator_u64(value) {
-                if iszero(eq(value, cleanup_u64(value))) { $Abort(95) }
-            }
 
             // Dummy calls to reference callables for Yul optimizer
             let $dummy0 := A2_M_fallback(sload(100))
@@ -145,6 +70,30 @@ object "A2_M" {
                 $result := x
             }
 
+            function abi_encode_tuple_$u64$(headStart ,value_0) -> tail {
+                tail := add(headStart, 32)
+                abi_encode_u64(value_0, add(headStart, 0))
+            }
+            function abi_encode_u64(value, pos) {
+                mstore(pos, cleanup_u64(value))
+            }
+            function cleanup_u64(value) -> cleaned {
+                cleaned := and(value, 0xffffffffffffffff)
+            }
+            function abi_decode_tuple_$u64$(headStart, dataEnd) -> value_0 {
+                if slt(sub(dataEnd, headStart), 32) { $Abort(96) }
+                {
+                    let offset := 0
+                    value_0 := abi_decode_u64(add(headStart, offset), dataEnd)
+                }
+            }
+            function abi_decode_u64(offset, end) -> value {
+                value := calldataload(offset)
+                validator_u64(value)
+            }
+            function validator_u64(value) {
+                if iszero(eq(value, cleanup_u64(value))) { $Abort(95) }
+            }
             function $Abort(code) {
                 revert(0, 0) // TODO: convention to store code
             }
@@ -154,3 +103,53 @@ object "A2_M" {
         }
     }
 }
+
+
+!! Move-To-Yul Diagnostics:
+ error: multiple #[receive] functions
+   ┌─ tests/fallback-receive-test/expect-failure/FallbackReciveFailure.move:10:2
+   │
+10 │ ╭     fun receive(x: u64): u64 {
+11 │ │         x
+12 │ │     }
+   │ ╰─────^
+
+error: receive function must be payable
+   ┌─ tests/fallback-receive-test/expect-failure/FallbackReciveFailure.move:10:2
+   │
+10 │ ╭     fun receive(x: u64): u64 {
+11 │ │         x
+12 │ │     }
+   │ ╰─────^
+
+error: receive function must not be a fallback or callable function
+   ┌─ tests/fallback-receive-test/expect-failure/FallbackReciveFailure.move:10:2
+   │
+10 │ ╭     fun receive(x: u64): u64 {
+11 │ │         x
+12 │ │     }
+   │ ╰─────^
+
+error: receive function must not have parameters
+   ┌─ tests/fallback-receive-test/expect-failure/FallbackReciveFailure.move:10:2
+   │
+10 │ ╭     fun receive(x: u64): u64 {
+11 │ │         x
+12 │ │     }
+   │ ╰─────^
+
+error: multiple #[fallback] functions
+   ┌─ tests/fallback-receive-test/expect-failure/FallbackReciveFailure.move:10:2
+   │
+10 │ ╭     fun receive(x: u64): u64 {
+11 │ │         x
+12 │ │     }
+   │ ╰─────^
+
+error: fallback function must not be a callable function
+   ┌─ tests/fallback-receive-test/expect-failure/FallbackReciveFailure.move:10:2
+   │
+10 │ ╭     fun receive(x: u64): u64 {
+11 │ │         x
+12 │ │     }
+   │ ╰─────^
